backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Background Tasks
celery==5.3.4
kombu==5.3.4

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.25.2
yfinance==0.2.28

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# File handling
python-magic==0.4.27
Pillow==10.1.0

# WebSocket support
websockets==12.0

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9

# Environment variables
python-dotenv==1.0.0